Update to v0.0.16

+++ lib/cybersource_rest_client/models/ptsv2credits_point_of_sale_information_emv.rb
@@ -37,7 +37,7 @@ module CyberSource
       {
         :'tags' => :'String',
         :'fallback' => :'BOOLEAN',
-        :'fallback_condition' => :'Float'
+        :'fallback_condition' => :'Integer'
       }
     end
