+++ lib/cybersource_rest_client/models/ptsv2payments_point_of_sale_information_emv.rb
@@ -44,10 +44,10 @@ module CyberSource
     def self.swagger_types
       {
         :'tags' => :'String',
-        :'cardholder_verification_method_used' => :'Float',
+        :'cardholder_verification_method_used' => :'Integer',
         :'card_sequence_number' => :'String',
         :'fallback' => :'BOOLEAN',
-        :'fallback_condition' => :'Float'
+        :'fallback_condition' => :'Integer'
       }
     end
